backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Job Queue
redis==5.0.1
rq==1.16.0

# Image Processing
Pillow==10.1.0

# Cloud Storage
boto3==1.34.10
cloudinary==1.36.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Validation
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
